pyproject.toml
[tool.poetry]
name = "fastapi-pomodoro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.112.2"}
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
redis = "^5.0.8"
pydantic-settings = "^2.4.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.32.3"
httpx = "^0.27.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
